Update nigeria budeshi spiders.
